+++ debian-linux/cli/material-for-mkdocs.sh
@@ -4,14 +4,13 @@
 # Python packages as tools (uv)
 # run as user (not root)
 
-if [ "$(whoami)" = "root" ]
-then
+if [ "$(whoami)" = "root" ]; then
  echo "Run the uv tool install script with your user account. Do not run as root or via sudo"
  exit
 fi
 
 echo "# ---------------------------------------"
 echo "Install Material for MkDocs as tool with supporting plugins"
-uv tool install --with mkdocs-material --with mkdocs-callouts --with mkdocs-glightbox --with mkdocs-git-revision-date-localized-plugin --with mkdocs-redirects --with mkdocs-rss-plugin --with pillow --with cairosvg mkdocs
+uv tool install mkdocs'<=1.9.9' --with mkdocs-material --with mkdocs-callouts --with mkdocs-glightbox --with mkdocs-git-revision-date-localized-plugin --with mkdocs-redirects --with mkdocs-rss-plugin --with pillow --with cairosvg
 echo "# ---------------------------------------"
 echo